LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/cssMathValue.tentative-expected.txt


PASS Constructing a CSSMathSum with no arguments throws a SyntaxError
PASS CSSMathSum can be constructed from a single number CSSUnitValue
PASS CSSMathSum can be constructed from more than one number CSSUnitValue
PASS CSSMathSum.operator is readonly
PASS Constructing a CSSMathProduct with no arguments throws a SyntaxError
PASS CSSMathProduct can be constructed from a single number CSSUnitValue
PASS CSSMathProduct can be constructed from more than one number CSSUnitValue
PASS CSSMathProduct.operator is readonly
PASS Constructing a CSSMathMin with no arguments throws a SyntaxError
PASS CSSMathMin can be constructed from a single number CSSUnitValue
PASS CSSMathMin can be constructed from more than one number CSSUnitValue
PASS CSSMathMin.operator is readonly
PASS Constructing a CSSMathMax with no arguments throws a SyntaxError
PASS CSSMathMax can be constructed from a single number CSSUnitValue
PASS CSSMathMax can be constructed from more than one number CSSUnitValue
PASS CSSMathMax.operator is readonly
PASS CSSMathNegate can be constructed from a single numberish value
PASS CSSMathNegate.operator is readonly
PASS CSSMathInvert can be constructed from a single numberish value
PASS CSSMathInvert.operator is readonly
PASS Constructing a CSSMathClamp with less than 3 arguments throws a SyntaxError
PASS Constructing a CSSMathClamp with different unit arguments throws a SyntaxError
PASS CSSMathClamp.operator is readonly
